+++ src/main/java/de/maxhenkel/peek/config/ClientConfig.java
@@ -9,6 +9,7 @@ public class ClientConfig {
     public final ConfigEntry<Boolean> peekChests;
     public final ConfigEntry<Boolean> peekBarrels;
     public final ConfigEntry<Boolean> peekDispensers;
+    public final ConfigEntry<Boolean> peekHoppers;
     public final ConfigEntry<Boolean> peekBeehives;
     public final ConfigEntry<Boolean> peekExplorationMaps;
     public final ConfigEntry<Boolean> peekCompasses;
@@ -19,6 +20,7 @@ public class ClientConfig {
         peekChests = builder.booleanEntry("peek_chests", true);
         peekBarrels = builder.booleanEntry("peek_barrels", true);
         peekDispensers = builder.booleanEntry("peek_dispensers", true);
+        peekHoppers = builder.booleanEntry("peek_hoppers", true);
         peekBeehives = builder.booleanEntry("peek_beehives", true);
         peekExplorationMaps = builder.booleanEntry("peek_exploration_maps", true);
         peekCompasses = builder.booleanEntry("peek_compasses", true);