+++ changelog.md
@@ -1,3 +1,4 @@
 - Added chest peeking
 - Added dispenser peeking
 - Added barrel peeking
+- Added hopper peeking